cleanup store needs and cometbft apis
