+++ server/v2/cometbft/types/peer.go
@@ -1,0 +1,6 @@
+package types
+
+import abci "github.com/cometbft/cometbft/abci/types"
+
+// PeerFilter responds to p2p filtering queries from Tendermint
+type PeerFilter func(info string) (*abci.ResponseQuery, error)
