+++ server/v2/cometbft/types/vote.go
@@ -1,0 +1,21 @@
+package types
+
+import (
+	"context"
+
+	abci "github.com/cometbft/cometbft/abci/types"
+	"google.golang.org/protobuf/proto"
+
+	corecomet "cosmossdk.io/core/comet"
+)
+
+// VoteExtensionHandler defines how to implement vote extension handlers
+type VoteExtensionsHandler interface {
+	ExtendVote(context.Context, *abci.RequestExtendVote) (*abci.ResponseExtendVote, error)
+	VerifyVoteExtension(context.Context, *abci.RequestVerifyVoteExtension) (*abci.ResponseVerifyVoteExtension, error)
+}
+
+type ConsensusInfo struct { // TODO: this is a mock, we need a proper proto.Message
+	proto.Message
+	corecomet.Info
+}
